fix(layout): M26.1 textarea 内容不再渲染（修复百度 CSS 泄漏）
严重渲染 bug：百度等大站把 CSS 文本塞进
<textarea id="s_is_result_css" style="display:none">
做延迟加载（JS 运行时再注入 <style>）。旧版把 textarea 当普通块级
元素，其 Text 子节点（字面是 CSS）被渲染出来，导致百度渲染输出 69%
是 CSS 噪音（3340 行里 2295 行 CSS），截图 50MB 且混乱。

根因定位（受控实验，不猜）：
- 标准 <style> 在 head/body/div 都不泄漏（黑名单正常）
- 但百度 CSS 在 <textarea> 里，绕过了 Element(style) 黑名单
- fetch 百度 HTML 确认：Element(textarea) > Text("<style>...CSS...")
- 受控实验 <textarea>SECRET</textarea> → SECRET 被渲染（复现）

修复：is_non_rendered_tag 黑名单加 'textarea'。真浏览器 textarea 内容是
表单初始值，不参与文档流渲染。

验证（受控实验 + 真实百度）：
- 受控：textarea 内容不再渲染，普通 <p> 正常 ✅
- 百度实测：3340→424 行（↓87%），2295→1 CSS 特征行（69%→0.2%）
- 截图内容干净：'百度首页 设置 登录'、'新闻 hao123 地图' 等
- 新增 integration_textarea.rs 2 e2e 锁定修复（防回归）

教训：之前说'百度 71% CSS 噪音 = 黑名单 bug'是错的，没做受控实验。
标准 <style> 黑名单一直工作，问题是 textarea 这种特殊 raw-text 容器。

workspace: 449 tests, 0 clippy warnings.

+++ crates/cli/src/screenshot.rs
@@ -222,7 +222,7 @@ mod tests {
             "'p' bottom ({bottom_p}) must be below 'b' bottom ({bottom_b}) — 否则 descender 丢失"
         );
         // 3. 字形都在 cell 内（不溢出到相邻行）
-        let line_height = (ascent + 4 + LINE_GAP as i32) as i32;
+        let line_height = ascent + 4 + LINE_GAP as i32;
         assert!(top_b >= 0, "'b' top ({top_b}) 不能为负");
         assert!(
             bottom_p < line_height,
